Fixing crash during commit when collection failed
If a source always fails to read, but others succeeded, the ensuing
commit of all collectors results in an unhandled IOError coming from
shutil.copy.  However, commit on an uncollected source should just
be a noop, as there is nothing to commit.

+++ os_collect_config/cache.py
@@ -60,4 +60,5 @@ def store(name, content):
 
 def commit(name):
     dest_path = get_path(name)
-    shutil.copy(dest_path, '%s.last' % dest_path)
+    if os.path.exists(dest_path):
+        shutil.copy(dest_path, '%s.last' % dest_path)
